add ex2 for cpp

+++ mpags-cipher.cpp
@@ -4,7 +4,7 @@ int main()
    std::cout<<"Hello World!\n";
 
    int a = 5;
-   int b =6.1;
+   double b =6.1;
    c =a*b;
    std::cout << c << std::endl;
